% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_view_campaign_adsets.R
\name{fb_view_campaign_adsets}
\alias{fb_view_campaign_adsets}
\title{View campaign adsets}
\usage{
fb_view_campaign_adsets(campaign_id, access_token)
}
\value{
Returns a tibble of variable dimensions, whereas rows are adsets and columns are id, name, age_max, age_min, genders, countries (and regions if so specified), configured_status, and effective_status.
}
\description{
This function looks up all adsets associated with a campaign and returns them as a tibble.
}
\examples{
fb_view_campaign_adsets(campaign_id = "123456789",
                                   access_token = "123456789")
}
\keyword{ad-sets}
\keyword{facebook}
