% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_age_into_quota_groups.R
\name{sort_age_into_quota_groups}
\alias{sort_age_into_quota_groups}
\title{Sort Age into Quota Groups}
\usage{
sort_age_into_quota_groups(age, age_brackets)
}
\arguments{
\item{age}{A respondent's age of type integer which is to be placed into a user-supplied age group.}

\item{age_brackets}{A dataframe with the following columns: age_group (character or factor), lower (numeric), and upper (numeric)}
}
\value{
Returns an age bracket. #???? not sure that's what it does
}
\description{
This function takes a user-supplied integer (age) and sorts it into the groups of the provided quota. This can be helpful if the quota aggregates the respondents' ages into age groups.
}
\examples{
sort_age_into_quota_groups(20, quota = us_quota) ### This might be the wrong quota
}
\keyword{categorization}
\keyword{cleaning}
\keyword{quota}
