% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_quota_age.R
\name{split_quota_age}
\alias{split_quota_age}
\title{Split Quota Age Groups into Separate Columns}
\usage{
split_quota_age(quota, age_variable = age)
}
\arguments{
\item{quota}{A dataframe with the following columns: location (character or factor), gender (character or factor), age group (character or factor), and population (numeric)}

\item{age_variable}{The name of the age_variable from which to extract the lower and upper bounds}
}
\value{
Returns the same dataframe as before but now with columns "lower" and "upper", which contain the bounds of the age variable.
}
\description{
This function takes a quota and the name of the age_group variable and splits the age group into two columns, one for the lower bound of the age group, one for the upper bound.
}
\examples{
split_quota_age(us_quota, age_variable = age)
}
\keyword{cleaning}
\keyword{quota}
