% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rel_freq.R
\name{get_rel_freq}
\alias{get_rel_freq}
\title{Get relative frequencies}
\usage{
get_rel_freq(quota, location = name, freq_raw = value)
}
\arguments{
\item{quota}{dataframe or tibble}

\item{location}{factor or character}

\item{freq_raw}{numeric}
}
\value{

}
\description{
This function takes a quota tibble or dataframe as its input and returns a tibble as its output.
The function assumes that one column, the "location" column, includes the location of your quota, e.g. a country, county, or city. This can also just be a single value for all rows.
The freq_raw argument encodes the raw number of people.
}
